Extract firefox check into environment.js
