+++ extension/src/html/background.html
@@ -21,6 +21,7 @@ limitations under the License.
     <script type="module" src="../js/main.js"></script>
     <script type="module" src="../js/companion.js"></script>
     <script type="module" src="../js/constants.js"></script>
+    <script type="module" src="../js/environment.js"></script>
     <script type="module" src="../js/logger.js"></script>
     <script type="module" src="../js/floater.js"></script>
     <script type="module" src="../js/notifier.js"></script>
